Remove clamping to 179.999
The clamping code was causing a number of headaches because it would
occasionally create very small sections of invalid/self-intersecting
geometry. CMR allows for values of 180 and -180 anyway, so it's actually
simpler to just remove the clamping. If we REALLY need the clamping
behavior for some reason, it would be better to implement it as a
separate transformation step after the polygons have been split.